requirements.txt
Django==2.0.3
graphene-django==2.0.0
dj-database-url==0.5.0
djangorestframework==3.7.7
markdown==2.6.11
django-filter==1.1.0
selenium==3.9.0
psycopg2-binary==2.7.4